scripts/requirements.txt
beautifulsoup4==4.12.2
chromedriver_autoinstaller==0.6.2
pymongo==3.12.0
python-dotenv==1.0.0
pytz==2022.1
requests==2.25.1
selenium==3.141.0
tqdm==4.65.1
